Move s3 util functions into utils folder
It's not actually used as a standalone script, so doesn't belong
in the scripts directory.
