+++ src/scripts/cronjobs/syncBreaches.ts
@@ -18,7 +18,7 @@ import {
   upsertBreaches,
   updateBreachFaviconUrl,
 } from "../../db/tables/breaches";
-import { uploadToS3 } from "../s3.js";
+import { uploadToS3 } from "../../utils/s3.js";
 
 const SENTRY_SLUG = "cron-sync-breaches";
